wheat prices
The monthly price of wheat (hard red winter) in the United States reached an all time high in May 2022, at over 520 U.S. dollars per metric ton. The unprecedented price increase began in mid-2020, due to the impact of the Covid-19 pandemic, and was later exacerbated by the Russo-Ukrainian War in March 2022. Before the war, Russia and Ukraine were among the world's five largest wheat exporters, and around one third of all international wheat imports came from these two countries.



The increase of 96 dollars per ton between February and March 2022 was the single largest price hike in U.S. history, and was only the second time that prices had exceeded 400 dollars - the first time this happened was due to the financial crisis of 2008. In the five years before the Covid-19 pandemic, the price of wheat generally fluctuated between 150 and 230 dollars per ton.